# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Api do gerenciamento de times. Usuário Admin para testes [ Email: admin@example.com | Senha: admin123 ] "
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/login": {
      "post": {
        "description": "Entrar na conta.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário logado!"
          },
          "400": {
            "description": "Dados incorretos, tente novamente."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "description": "Sair da conta.",
        "responses": {
          "200": {
            "description": "Usuário deslogado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/install": {
      "get": {
        "description": "Preenche e instala o banco de dados.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/projects/create": {
      "post": {
        "tags": ["Projects"],
        "description": "Cria um novo projeto.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Nome do Projeto"
                },
                "description": {
                  "type": "string",
                  "example": "Descrição do Projeto"
                },
                "teamId": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso!"
          },
          "400": {
            "description": "Dados inválidos ou time não encontrado."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/projects/read": {
      "get": {
        "tags": ["Projects"],
        "description": "Busca todos os projetos com paginação.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Número máximo de projetos por página."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Número da página a ser retornada."
          }
        ],
        "responses": {
          "200": {
            "description": "Projetos encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Projetos encontrados!"
                    },
                    "projects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "teamId": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados de paginação inválidos."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/projects/update/{id}": {
      "put": {
        "tags": ["Projects"],
        "description": "Atualiza os dados de um projeto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto a ser atualizado."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Novo Nome do Projeto"
                },
                "description": {
                  "type": "string",
                  "example": "Nova Descrição do Projeto"
                },
                "teamId": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou time não encontrado."
          },
          "404": {
            "description": "Projeto não encontrado."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/projects/my-projects": {
      "get": {
        "tags": ["Projects"],
        "description": "Busca os projetos pertencentes ao usuário autenticado.",
        "responses": {
          "200": {
            "description": "Projetos pertencentes encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Projetos pertencentes encontrados."
                    },
                    "projects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "teamId": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/projects/delete/{projectId}": {
      "delete": {
        "tags": ["Projects"],
        "description": "Deleta um projeto específico.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto a ser deletado."
          }
        ],
        "responses": {
          "204": {
            "description": "Projeto deletado com sucesso."
          },
          "403": {
            "description": "Usuário não tem permissão para deletar este projeto."
          },
          "404": {
            "description": "Projeto ou time associado não encontrado."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    }
  }
}